+++ public_html/index.php
@@ -13,6 +13,10 @@
 
 require_once '../lib-common.php';
 
+// Tell search engines not to index, may use excessive requests
+$outputHandle = outputHandler::getInstance();
+$outputHandle->addMeta('name', 'ROBOTS', 'noindex,nofollow');
+
 COM_setArgNames(array('mode', 'asin'));
 $mode = COM_getArgument('mode');
 $asin = COM_getArgument('asin');